+++ pkg/lsp/server.go
@@ -27,6 +27,8 @@ type Handler interface {
 	TextDocumentCompletion(params CompletionParams) ([]CompletionItem, error)
 	TextDocumentHover(params HoverParams) (*Hover, error)
 	TextDocumentSemanticTokensFull(params SemanticTokensParams) (*SemanticTokens, error)
+	TextDocumentFormatting(params DocumentFormattingParams) ([]TextEdit, error)
+	TextDocumentRangeFormatting(params DocumentRangeFormattingParams) ([]TextEdit, error)
 	WorkspaceDidChangeConfiguration(params DidChangeConfigurationParams) error
 }
 
@@ -187,12 +189,36 @@ func (s *Server) handleRequest(req *Request) error {
 	case "shutdown":
 		return s.sendResponse(req.ID, nil)
 
+	case "textDocument/formatting":
+		var params DocumentFormattingParams
+		if err := json.Unmarshal(req.Params, &params); err != nil {
+			return s.sendErrorResponse(req.ID, InvalidParams, "Invalid params")
+		}
+
+		result, err := s.handler.TextDocumentFormatting(params)
+		if err != nil {
+			return s.sendErrorResponse(req.ID, InternalError, err.Error())
+		}
+
+		return s.sendResponse(req.ID, result)
+
+	case "textDocument/rangeFormatting":
+		var params DocumentRangeFormattingParams
+		if err := json.Unmarshal(req.Params, &params); err != nil {
+			return s.sendErrorResponse(req.ID, InvalidParams, "Invalid params")
+		}
+
+		result, err := s.handler.TextDocumentRangeFormatting(params)
+		if err != nil {
+			return s.sendErrorResponse(req.ID, InternalError, err.Error())
+		}
+
+		return s.sendResponse(req.ID, result)
+
 	// Optional capabilities - respond with null to indicate not supported
 	case "textDocument/documentSymbol",
 		"textDocument/definition",
 		"textDocument/references",
-		"textDocument/formatting",
-		"textDocument/rangeFormatting",
 		"textDocument/onTypeFormatting",
 		"textDocument/codeAction",
 		"textDocument/codeLens",